More for base
Bathrooms done pretty much, armory area has more.  The colliders in the
stage are not functioning off of each wall.  Instead there are fewer
invisible colliders that stretch to fit where they are so that it just
makes a complete wall.  Some of these colliders have not been fixed as
the level was being placed.  Next update is having the colliders in
proper place